Seperated addItem dom code from dom
